Add FormatUtils class.
